remove obsolete prep_roster_sync script (Recidiviz/recidiviz-data#48578)
## Description of the change

This script was originally introduced to prepare states that had data in
the Roster table for roster sync by:
* copying their data over to the UserOverride table
* removing users that hadn't been active in a long time
* updating some data to match what roster sync would produce

Now that roster sync is enabled in all states that had data in the
Roster table pre-roster-sync, there are no users left that we need to
attempt to preserve. Since the Roster table is now only updated by
roster sync, turning on roster sync in a state won't remove anyone's
access or change their attributes if they were already in the admin
panel.

Tagging in @alexabatino and @Gina-Valderrama as per
https://www.notion.so/recidiviz/Polaris-POCs-21b7889f4d1980eea689e28b3e84e03d

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b5eea463bfd8cd5ed5de0b466ecfc58c34a9e4b0