[US_OR] Delete all views referencing legacy sentences tables (Recidiviz/recidiviz-data#48587)
## Description of the change

This PR deletes any OR BQ view that references
`state_supervision_sentence`, `state_incarceration_sentence` or
`sessions.sentences_preprocessed` (which, in turn, references those
tables). It also deletes a handful of views that become unused once all
those views are deleted. We need to do this in order to fully delete v1
sentences infra. If we re-establish a relationship with OR and start
getting data again, we should ingest the new sentencing data into the v2
sentences schema.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#46263

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0b62f9d0d693f13edf72fcdeba68736268445f51