[US_AZ] Ingest STATE_VIOLATION_ID as incident_metadata (Recidiviz/recidiviz-data#49914)
## Description of the change

This will make it easier to attach changes in earned time to violations
in the AZ JII app, if and when we choose to do that!

See
[here](https://www.notion.so/recidiviz/Earned-Time-Credits-in-AZ-21b7889f4d1980e98346e25a3649e964?source=copy_linkRecidiviz/recidiviz-data#2497889f4d1980258768e508a3f677b5)
for details.

[sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-09-24_06_43_27-5973185079412228472;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
[Recidiviz/recidiviz-data#46164](https://github.com/Recidiviz/recidiviz-data/issues/46164)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 01f8c135f942b9b4a4933d351aedf9aef9b1c275