[Insights] fix MI split parole/probation metrics by using attrs.evolve (Recidiviz/recidiviz-data#49955)
## Description of the change

This fixes the weird bug that was discovered when @Gina-Valderrama tried
to test out the MI split parole/probation metrics by setting her FV in
auth0 and loading the page in staging.

The main problem was here:
https://github.com/Recidiviz/recidiviz-data/blob/c01d8e8b5e72981c5dfce0802dd8890d61fa9a83/recidiviz/outliers/querier/querier.py#L1965-L1967.

This part of the code was setting `backend_config_obj` as an actual
reference to the same object as `self.get_outliers_backend_config()`, so
when `backend_config_obj.metrics = self.get_outcomes_metrics()` was run,
it changed the actual value of self.get_outliers_backend_config() for
the whole backend. This is why it worked when Gina set her feature
variants but then failed for everyone else afterward: the metrics list
was getting filtered to only include the split metrics for _everyone_,
so then when people loaded the page without the FV to include split
metrics set for them, it showed they didn't have access to any metrics
for MI. This also explains why a staging deploy fixed it.

This PR fixes the bug by using attrs.evolve, and also sets `frozen=True`
on all the attrs classes in types.py since we probably don't expect any
of them to be mutable, and it's safer to just set them all this way for
now and remove it if we have a real use case later.

Confirmed that by making only the `frozen` change, current tests failed
(so we would have caught the issue had the classes been marked as
frozen). Tested the fix by running the FE and the case triage backend
locally and refreshing the page both with and without the FV set in
auth0 (which did not work before this change)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/39162

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 280d5dfde28e9da61cd088bd7b1205c11a8745e2