[lineage] remove concept of selected node (Recidiviz/recidiviz-data#49947)
## Description of the change

this PR removes the concept of a selected node -- we had been using it
as a way to anchor the graph, but that wasn't necessarily clear to the
user. instead, we parse the query params `?start_urn=<urn>` in order to
allow there to be direct links to nodes while still being able to remove
that node throughout the contract/expand/filter process.

we had used the selected node for a few things -- first among them was
to avoid having to try to determine which graph to pick if there were
any disconnected subgraphs after a filter way applied; in it's place we
just pick the subgraph that is largest

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d394048c2565afe8b78ab9497aabdac3925883f4