[US_AZ] Refactored Mandatory Literacy criteria to have enrollment date inside its reasons blobs (Recidiviz/recidiviz-data#47737)
## Description of the change

We want to show in the TPR side panel when a resident is enrolled in
mandatory literacy beyond the current state, where we show it in the
Additional Information tab. More specifically, we want to show it in the
criteria line as "Does not meet functional literacy BUT actively
enrolled in functional literacy". In order to do so, the enrollment
status needed to be involved inside the reasons blob of the functional
literacy criteria without affecting eligibility. This was done by taking
our current enrollment criteria and changing it to be used for TPR and
DTP, keeping DTP unchanged (as enrollment is equivalent to meeting the
criteria) at the same time.

This has been uploaded to sandbox under `shalin` with all validations
passing and eligibility completely unaffected, as we were hoping. We can
now see the enrollment date in the reasons blob for TPR mandatory
literacy criteria.

Query for loading to sandbox
```
python -m recidiviz.tools.load_views_to_sandbox \
        --sandbox_dataset_prefix shalin --state_code_filter US_AZ --prompt manual \
        --dataset_ids_to_load validation_views \
	      --view_ids_to_load workflows_views.us_az_approaching_acis_or_recidiviz_tpr_request_record,workflows_views.us_az_approaching_acis_or_recidiviz_dtp_request_record,task_eligibility_spans_us_az.overdue_for_recidiviz_tpr_request,task_eligibility_spans_us_az.overdue_for_recidiviz_dtp_request,task_eligibility_criteria_us_az.enrolled_in_functional_literacy_tpr,task_eligibility_criteria_us_az.enrolled_in_functional_literacy_dtp,task_eligibility_criteria_us_az.enrolled_in_or_meets_mandatory_literacy_dtp,task_eligibility_criteria_us_az.enrolled_in_or_meets_mandatory_literacy_tpr \
        --update_ancestors False \
        --update_descendants True
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#47660

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 34fa118c3f7675ea8089d6bce47cde2ffc2f1ced