[Workflows] Create external request endpoint for emailing the logged-in user (Recidiviz/recidiviz-data#48593)
## Description of the change

Make a new endpoint `POST
/workflows/external_request/<state>/email_user` that can be used to send
an email to the currently authenticated user. This is only enabled in
Texas and will be used for the TX Tasks Home Contact Route Planner, so
that a user can email themself a Google Maps link to the route.

We decided against allowing the user to trigger an email to an arbitrary
address to ensure we'll only be emailing people who actually have
Recidiviz accounts. If they really want to get the email to another
address (e.g. a personal email or a coworker), they could just forward
it.

#### Testing

Unit tests have been added. There are other tickets I'd like to do
before implementing the button to email a route on the frontend, so this
won't be end-to-end-testable for a bit.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#46907

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: df4dbc7eaefdd2e44605172aa4e060275720583c