[Pathways][US_NY] Enable NY Pathways in product_views and staging exports (Recidiviz/recidiviz-data#50062)
## Description of the change

Enables `US_NY` as a pathways enabled state to hydrate Pathways data for
NY. Also enables the metric export for NY in staging. I will change this
to production once we have location data in prod.

I *think that this is all I need to do to get the NY cloudSQL db created
and populated, but @danawillow I would appreciate a second opinion on
this. In `pathways_database_manager` it pulls from
`pathways_enabled_states` so I believe that the db will just be created
without any additional config. I didn't see any migrations for the other
Pathways states, but please do let me know if there's something else I
need to do.

I ran the product views into a sandbox
`jovergaag_ny_pathways_dashboard_views` and the data is mostly hydrated.
We're missing a couple of dimensions but those aren't needed for the v0.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#50061

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c2a08b7c1f7bf3f02a5131b8e4955ca93c60290c