[Auth] Clean up user overrides after roster sync (Recidiviz/recidiviz-data#44320)
## Description of the change

Adds a call to the `cleanup_user_overrides` script in the
`import_ingested_users` endpoint to automate UserOverride cleanup after
roster sync.

I was thinking through whether `cleanup_user_overrides` should also be
called anywhere else (e.g., when UserOverride is updated via the admin
panel), but I think having it only here is fine. Because not everyone
has a Roster record (e.g., users added via the admin panel) and for
those who do, the Roster sort of acts as the default/source of truth, I
think it makes sense to only do cleanup after we have updated Roster
information.

I also added a test--not sure it's strictly necessary since the function
is tested separately in
https://github.com/Recidiviz/recidiviz-data/blob/main/recidiviz/tests/tools/auth/cleanup_user_overrides_test.py,
but I wanted to check that it's working properly within the endpoint. We
have to test that with a call to the db because GET /users will return a
combo of Roster and UserOverride data, and we wouldn't be able to see
the specific changes to UserOverride.

One issue though is that this is failing the
`validate_source_visibility` check--what's the best way to remedy that?
Maybe moving `cleanup_user_overrides` to a different folder, or is it
okay to add it to the valid modules list? I'm not sure if we want to
keep it as a standalone script to use on an ad hoc basis in addition to
running automatically after roster sync.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#43142

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c6cbd3aacb42dc1953aa95b131fa5610e81b244f