Add ingest claude.md (Recidiviz/recidiviz-data#49934)
## Description of the change
Add's a CLAUDE.md about the ingest process specifically. Claude
generated this itself. I think it's a fine baseline -- IMO we should
commit and adjust as needed, but any small suggestions or call-outs of
down-right wrong information are of course super welcome!

By including this in the `ingest` directory, CLAUDE will only load this
context when it has to read another file in the `ingest` directory. More
details
[here](https://docs.claude.com/en/docs/claude-code/memory#how-claude-looks-up-memories).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1b934890db44bf58198c095c87ef8f64cddb6ef3