[US_TX][Workflows] Add DPO addresses to supervision staff record (Recidiviz/recidiviz-data#50119)
## Description of the change

Add a new field, `dpo_address`, to the state-specific data in the
supervision staff record for staff in Texas. This will help fill in the
logged-in user's office address automatically for the Tasks Home Contact
Route Planner project. The field has the same shape as
`current_physical_residence_address_structured` in the client record.

The key change to get office addresses from state raw data is in
`reference_views.us_tx_location_metadata`. I had to make additional
changes to get a struct to propagate properly through
`sessions.supervision_staff_attribute_sessions` to the staff record -
thanks to Mayuka for helping with the approach here!

This PR takes a different approach from the one originally outlined in
the tickets/design doc because I realized that adding DPO information to
the location record doesn't really make sense - the location record is
currently only for locations that are associated in some way with a
client or resident, like a facility someone's housed at or eligible for
release to. So it made more sense to me to add the addresses to the data
associated with each staff member, which already exists.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#46902
Closes Recidiviz/recidiviz-data#46903

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3eb1fe5521d8bdddc1672895bbd962c5a0f47c4c